refactor: mv runStakeKit.js to stakeKit.js
